Move override_centos.py to _internal
