+++ certbot-apache/certbot_apache/tests/centos6_test.py
@@ -5,7 +5,7 @@ from certbot.compat import os
 from certbot.errors import MisconfigurationError
 
 from certbot_apache import obj
-from certbot_apache import override_centos
+from certbot_apache._internal import override_centos
 from certbot_apache import parser
 from certbot_apache.tests import util
