+++ certbot-apache/certbot_apache/tests/centos_test.py
@@ -8,7 +8,7 @@ from certbot.compat import filesystem
 from certbot.compat import os
 
 from certbot_apache import obj
-from certbot_apache import override_centos
+from certbot_apache._internal import override_centos
 from certbot_apache.tests import util
